% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_radelement_cdes.R
\name{get_radelement_cdes}
\alias{get_radelement_cdes}
\title{GET response to a code. This function will determine the vocabulary system automatically.}
\usage{
get_radelement_cdes(search = NULL)
}
\arguments{
\item{code}{one of RADLEX, LOINC, DICOM, SNOMEDCT}

\item{elements}{TRUE if elements are desired.}
}
\description{
GET response to a code. This function will determine the vocabulary system automatically.
}
