% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_radelement_code.R
\name{query_radelement_code}
\alias{query_radelement_code}
\title{GET response to a code. This function will determine the vocabulary system automatically.}
\usage{
query_radelement_code(code, values = FALSE, elements = TRUE, page = 1)
}
\arguments{
\item{code}{one of RADLEX, LOINC, DICOM, SNOMEDCT}

\item{elements}{TRUE if elements are desired.}
}
\description{
GET response to a code. This function will determine the vocabulary system automatically.
}
